Zenith angle functionality
Added the beginning of functionality to handle a grid with multiple zenith angles. This is not yet implemented (only a single ZA is in the grid now) because the ZA needs to be calculated per observation, but I wanted to add the basic functionality.

Also, the output model only has two entries: combined line absorption and continuum absorption.