Add zenith angle calculation
Added zenith angle calculation for HPF targets, because the header altitude is for the HET primary (fixed) but targets can have a variety of zenith angles along their observation track.

The calculation uses astropy coordinates, and first tries to use queue coordinates for the target. If there are not values for the QRA and QDEC header keys, then it defaults to using the TCS RA/Dec. Sometimes those are also empty, at which it falls back to the fixed altitude of the HET primary.