Add generalized model interpolation
Changed the model interpolation functions to be generalized for any species. The interpolation works across zenith angle and a generic "abundance" that can have different formats for different species. Currently set to only work for the molecules in the HPF bandpass.

Also had to change the function that parsed the model grid files.